Updated emails.py Created billing.py
emails.py now has the capability to choose a random charity to put at the end of each email

billing.py will have functions to verify if a users bank account has enough money to make a purchase/donation. This will most likely return a bool for if the payment goes through or not. THIS IS ALL SIMULATED WITH NO BANK API